Simplify argument builder logic in toolhelper module.
The simplification now requires each tool passes an explicit list of all
options that it cares about instead of depending on regex.  This change
requires more work upfront in each tool, but removes tons of
boilerplate, including parsing the entire possible list of
options--including the ones a tool doesn't care about--multiple times.

The changes here only work for the cmake build tool and only the dry-run
executor was tested; there's a good chance there are bugs actually
trying to run CMake (the separator is hard-coded--and wrong), and tools
like git simply won't work.